moving tests around to the right folders to solve cycles; fela-bindings is now directly tested as integration
