+++ packages/fela-enzyme/package.json
@@ -28,9 +28,9 @@
     "react": "*"
   },
   "devDependencies": {
-    "enzyme": "3.6.0",
-    "enzyme-to-json": "3.3.4",
-    "enzyme-adapter-react-16": "1.5.0"
+    "enzyme": "3.3.0",
+    "enzyme-to-json": "3.3.0",
+    "enzyme-adapter-react-16": "1.7.0"
   },
   "dependencies": {
     "cssbeautify": "^0.3.1",
