+++ packages/fela-integration/package.json
@@ -9,8 +9,11 @@
     "fela": "^6.2.3",
     "fela-monolithic": "^5.0.28",
     "fela-tools": "^5.2.3",
+    "raf": "^3.4.1",
     "react": "^16.6.0",
+    "react-dom": "^16.6.1",
     "react-fela": "^8.0.4",
-    "react-test-renderer": "^16.6.0"
+    "react-test-renderer": "^16.6.0",
+    "jest-react-fela": "^1.0.1"
   }
 }
