+++ packages/fela-integration/src/jest-react-fela_react-fela/__tests__/ThemeProviderFactory-test.js
@@ -1,0 +1,18 @@
+import 'raf/polyfill'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+
+import { createSnapshot } from 'jest-react-fela'
+import { ThemeProvider, FelaTheme } from 'react-fela'
+
+describe('Using the ThemeProvider', () => {
+  it('should pass the theme to rule props', () => {
+    const snapshot = createSnapshot(
+      <ThemeProvider theme={{ color: 'red' }}>
+        <FelaTheme>{theme => <div>{JSON.stringify(theme)}</div>}</FelaTheme>
+      </ThemeProvider>
+    )
+
+    expect(snapshot).toMatchSnapshot()
+  })
+})
